requirements.txt
aiobotocore==2.22.0
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.16.1
alpaca-py==0.40.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-generator==1.10
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
botocore==1.37.3
cachetools==5.5.2
ccxt==4.4.88
certifi==2025.4.26
certipy==0.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
concurrent-log-handler==0.9.27
contourpy==1.3.0
cryptography==45.0.3
cssselect==1.3.0
curl_cffi==0.11.2
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docstring_parser==0.16
duckdb==1.3.0
EditorConfig==0.17.0
exceptiongroup==1.3.0
exchange_calendars==4.5.6
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fire==0.7.0
fonttools==4.58.1
fqdn==1.5.1
fredapi==0.5.2
freezegun==1.5.2
frozendict==2.4.6
frozenlist==1.6.2
fsspec==2025.5.1
ghp-import==2.1.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsbeautifier==1.15.4
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterhub==5.3.0
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaleido==0.2.1
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
lxml==5.4.0
lxml_html_clean==0.4.2
mainsequence @ file:///opt/code/mainsequence-sdk
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mergedeep==1.3.4
minio==7.2.15
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-mermaid2-plugin==1.2.1
msgpack==1.1.0
multidict==6.4.4
multitasking==0.0.11
narwhals==1.41.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
notebook_shim==0.2.4
numexpr==2.10.2
numpy==2.0.2
oauthlib==3.2.2
opentelemetry-api==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
opentelemetry-exporter-otlp-proto-http==1.16.0
opentelemetry-proto==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
overrides==7.7.0
packaging==25.0
pamela==1.2.0
pandas==2.3.0
pandas_market_calendars==5.1.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
polygon-api-client==1.14.6
portalocker==3.1.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.8.0
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyluach==2.2.0
pymdown-extensions==10.15
pyparsing==3.2.3
pytest==8.4.0
pytest-freezegun==0.4.2
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-logstash==0.4.8
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==26.4.0
QuantStats==0.0.64
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
s3fs==2025.5.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.46.2
structlog==25.4.0
tabulate==0.9.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tomli==2.2.1
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.61
zipp==3.22.0
